+++ tools/make_donate_qr.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (C) 2026 LiuFudi
 #
-# This file is part of fn-fancontrol, licensed under the GNU General Public
+# This file is part of niufan, licensed under the GNU General Public
 # License version 3 or (at your option) any later version.
 # See the LICENSE file for the full text.
 """Compile the donation QR images into a JavaScript module.
@@ -44,7 +44,7 @@ HEADER = """\
 /* SPDX-License-Identifier: GPL-3.0-or-later
  * Copyright (C) 2026 LiuFudi
  *
- * This file is part of fn-fancontrol, licensed under the GNU General Public
+ * This file is part of niufan, licensed under the GNU General Public
  * License version 3 or (at your option) any later version.
  * See the LICENSE file for the full text.
  *